+++ runpod/api/mutations/templates.py
@@ -1,11 +1,12 @@
 """ RunPod | API Wrapper | Mutations | Templates """
 
-# pylint: disable=too-many-arguments
+# pylint: disable=too-many-arguments, too-many-branches
+
 
 def generate_pod_template(
-        name:str, image_name:str, docker_start_cmd:str=None,
-        container_disk_in_gb:int=10, volume_in_gb:int=None, volume_mount_path:str=None,
-        ports:str=None, env:dict=None, is_serverless:bool=False
+        name: str, image_name: str, docker_start_cmd: str = None,
+        container_disk_in_gb: int = 10, volume_in_gb: int = None, volume_mount_path: str = None,
+        ports: str = None, env: dict = None, is_serverless: bool = False, registry_auth: str = None
 ):
     """ Generate a string for a GraphQL mutation to create a new pod template. """
     input_fields = []
@@ -44,11 +45,15 @@ def generate_pod_template(
     else:
         input_fields.append('env: []')
 
-
     if is_serverless:
         input_fields.append('isServerless: true')
     else:
         input_fields.append('isServerless: false')
+
+    if registry_auth is not None:
+        input_fields.append(f'containerRegistryAuthId : "{registry_auth}"')
+    else:
+        input_fields.append('containerRegistryAuthId : ""')
 
     # ------------------------------ Enforced Fields ----------------------------- #
     input_fields.append('startSsh: true')